Move SDL mappigs to the separate file
